% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_expected_p_uniform.R
\name{calc_expected_p_uniform}
\alias{calc_expected_p_uniform}
\title{Calculate Expected P-values from Uniform}
\usage{
calc_expected_p_uniform(observed_p, na.rm = FALSE)
}
\arguments{
\item{observed_p}{A numeric vector of observed p-values. Must be non-negative.}

\item{na.rm}{Logical; if TRUE, any NA values in `observed_p` will be removed before computation.
Default is FALSE, which will issue a warning if NAs are present but not remove them.}
}
\value{
Returns a numeric vector of the same length as `observed_p` (after NA removal, if applicable), 
        containing the expected uniform p-values corresponding to the ranks of the observed p-values.
}
\description{
This function computes the expected uniform p-values based on the observed p-values. 
It can optionally remove NA values before computation. The function is designed 
to transform a set of observed p-values such that they follow a uniform distribution, 
which is useful for certain statistical analyses where uniformity of p-values is expected under the null hypothesis.
}
\details{
The function first checks that `observed_p` is a numeric vector and contains 
         non-negative values. It then handles NA values according to the `na.rm` parameter. 
         The expected uniform p-values are calculated by ranking the observed p-values and 
         then scaling these ranks linearly between 1/n and 1, where n is the number of 
         observed p-values after NA handling. This transformation assumes that the input p-values 
         are drawn from a continuous uniform distribution.
}
\examples{
# Example without NAs and na.rm = FALSE
observed_p <- c(0.05, 0.2, 0.01, 0.5, 0.95)
calc_expected_p_uniform(observed_p)

# Example with NAs and na.rm = TRUE
observed_p_with_na <- c(0.05, NA, 0.01, 0.5, 0.95)
calc_expected_p_uniform(observed_p_with_na, na.rm = TRUE)

}
