% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_f_from_theta.R
\name{calc_f_from_theta}
\alias{calc_f_from_theta}
\title{Calculate Inbreeding Coefficient from Theta}
\usage{
calc_f_from_theta(pA, theta)
}
\arguments{
\item{p}{Numeric, the frequency of the minor allele in the population.}

\item{f}{Numeric, the inbreeding coefficient.}
}
\value{
Numeric, the inbreeding coefficient `f` recalculated from `theta` and `pA`.
}
\description{
Computes the inbreeding coefficient `f` based on the theta parameter used in
Hardy-Weinberg equilibrium tests and allele frequency `p`.
}
\examples{
pA <- 0.2  # Minor allele frequency
f <- 0.01 
calc_f_from_theta(pA, f)

}
