% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_log_norm_c.R
\name{calc_log_norm_c}
\alias{calc_log_norm_c}
\title{Calculate Log of the Normalizing Constant C}
\usage{
calc_log_norm_c(N, nA, nB, theta)
}
\arguments{
\item{N}{The total number of individuals in the population.}

\item{nA}{The total number of minor allele A in the population.}

\item{nB}{The total number of common allele B in the population. Calculated as 2*N - nA.}

\item{theta}{The theta parameter reflecting the relationship between heterozygotes and homozygotes frequencies.}
}
\value{
The log of the normalizing constant C, used in the computation of probabilities.
}
\description{
This function calculates the log of the normalizing constant C, used in the computation
of probabilities for genetic heterozygote configurations under a given theta parameter.
It applies the log-sum-exp trick to maintain numerical stability.
}
\examples{
N <- 100
nA <- 21
nB <- 2 * N - nA
theta <- 4
calc_log_norm_c(N, nA, nB, theta)

}
