% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_log_p_hets.R
\name{calc_log_p_hets}
\alias{calc_log_p_hets}
\title{Calculate Log Probability of nAB Heterozygotes}
\usage{
calc_log_p_hets(N, nA, nAB, theta)
}
\arguments{
\item{N}{The total number of diploid individuals in the population.}

\item{nA}{The total number of copies of the minor allele A in the population.}

\item{nAB}{The observed number of heterozygotes in the population.}

\item{theta}{The theta parameter, which reflects the relationship between the frequencies of
heterozygotes and homozygotes. It's used to adjust the expected probabilities under
the Hardy-Weinberg equilibrium or other genetic models.}
}
\value{
The log of the probability of observing exactly nAB heterozygotes given N, nA, and theta.
}
\description{
Computes the logarithm of the probability of observing a specific number of heterozygotes (nAB)
in a population, given the total number of individuals (N), the total number of minor alleles (nA),
and a theta parameter that reflects the expected heterozygosity.
}
\details{
This function first calculates the log of the numerator, which includes the term for
heterozygotes weighted by the theta parameter, and the factorial of the total number of individuals (N).
It then subtracts the log of the denominator, which includes the factorials of the number of AA homozygotes,
AB heterozygotes, and BB homozygotes. Finally, it subtracts the log of the normalizing constant C, calculated
by `calc_log_normalizing_constant`, to obtain the final log probability.

The function ensures that the calculated number of AA and BB homozygotes are integers, as required for
valid configurations in a diploid population.
}
\examples{
N <- 100
nA <- 21
nAB <- 17
theta <- 4
calc_log_p_nAB_heterozygotes(N, nA, nAB, theta)

}
