% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_norm_c.R
\name{calc_norm_c}
\alias{calc_norm_c}
\title{Calculate Normalizing Constant}
\usage{
calc_norm_c(N, nA, nB, theta)
}
\arguments{
\item{N}{The total number of individuals in the population.}

\item{nA}{The number of alleles A in the population.}

\item{nB}{The number of alleles B in the population.}

\item{theta}{The inbreeding coefficient.}
}
\value{
The normalizing constant \( C \).
}
\description{
Computes the normalizing constant \( C \) given the population size \( N \), the number of alleles \( nA \), \( nB \), and inbreeding coefficient \( theta \).
}
\examples{
calc_norm_c(100, 40, 160, 0.01)
}
