% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_theta.R
\name{calc_theta}
\alias{calc_theta}
\title{Calculate Theta}
\usage{
calc_theta(N, K, M)
}
\arguments{
\item{N}{The total number of samples in the population. This should be a positive integer.}

\item{K}{The count of the minor allele in the population. This should be a positive integer less than or equal to 2*N.}

\item{M}{The number of homozygous carriers of the minor allele in the population. This should be a positive integer less than or equal to N.}
}
\value{
Returns the theta value as a numeric value. This value is a measure of the balance between
        the frequencies of the homozygous major, heterozygous, and homozygous minor genotypes.
}
\description{
This function calculates the theta value based on allele counts. Theta is calculated using
the formula \eqn{\theta = \frac{p_{AB}^2}{p_{AA} \cdot p_{BB}}}, where \eqn{p_{AA}}, \eqn{p_{AB}}, and
\eqn{p_{BB}} are the frequencies of the homozygous for the major allele, heterozygous, and homozygous for the
minor allele genotypes, respectively. These frequencies are derived from allele counts obtained
for a given population.
}
\examples{
N <- 10000 # samples in population
K <- 1000 # minor allele count
M <- 10 # homozygous carriers
theta <- calc_theta(N, K, M)

}
