% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_theta_from_f.R
\name{calc_theta_from_f}
\alias{calc_theta_from_f}
\title{Calculate Theta using Inbreeding Coeffecient}
\usage{
calc_theta_from_f(pA, f)
}
\arguments{
\item{pA}{Numeric, the frequency of the minor allele in the population.}

\item{f}{Numeric, the inbreeding coefficient.}
}
\value{
Numeric, the calculated theta value.
}
\description{
Computes the theta parameter used in Hardy-Weinberg equilibrium tests,
taking into account the inbreeding coefficient.
}
\examples{
pA <- 0.2  # Minor allele frequency
f <- 0.01 # Inbreeding coefficient
theta <- calc_theta_from_f(pA, f)
print(theta)

}
