% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_allele_counts}
\alias{get_allele_counts}
\title{Get Allele Counts}
\usage{
get_allele_counts(N, K, M)
}
\arguments{
\item{N}{Numeric, the total number of individuals in the population.}

\item{K}{Numeric, the total number of mutated gene copies observed (equivalent to minor allele count).}

\item{M}{Numeric, the number of samples that are homozygous carriers for the minor allele.}
}
\description{
This function calculates the allele counts based on the total number of individuals in the population (N), 
the total number of mutated gene copies observed (K), and the number of samples that are homozygous 
carriers for the minor allele (M). It returns a vector with the counts of homozygous alternate, 
heterozygous, and homozygous reference allele carriers, respectively.
}
