% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_mac}
\alias{get_mac}
\title{Get Minimum Allele Count (MAC)}
\usage{
get_mac(N, K, M)
}
\arguments{
\item{N}{Numeric, the total number of individuals in the population.}

\item{K}{Numeric, the total number of mutated gene copies observed (equivalent to minor allele count).}

\item{M}{Numeric, the number of samples that are homozygous carriers for the minor allele.}
}
\description{
This function calculates the minimum allele count (MAC) for a given set of genotype counts. It uses
the total number of individuals (N), the total number of mutated gene copies observed (K), and the number
of samples that are homozygous carriers for the minor allele (M) to determine the MAC.
}
