% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hwe_exact_test.R
\name{hwe_exact_test}
\alias{hwe_exact_test}
\title{Perform an Exact Test under Hardy-Weinberg Equilibrium}
\usage{
hwe_exact_test(
  N,
  K,
  M,
  theta = 4,
  alternative = "less",
  use_mid_p = TRUE,
  debug = FALSE
)
}
\arguments{
\item{N}{Numeric, the total number of individuals in the population.}

\item{K}{Numeric, the total number of mutated gene copies observed (equivalent to minor allele count)}

\item{M}{Numeric, the number of samples that have are homozygous carriers}

\item{theta}{Numeric, the inbreeding coefficient parameter; defaults to 4. This parameter reflects 
the degree of deviation from random mating.}

\item{alternative}{Character, specifies the alternative hypothesis and must be one of "less", 
"greater", or "two.sided". "Less" tests for a deficit of heterozygotes (indicating excess homozygosity),
and "greater" for an excess of heterozygotes.}

\item{use_mid_p}{Logical, indicating whether to use the mid-p correction for the p-value calculation in one-sided tests; defaults to TRUE.}

\item{debug}{Logical, if TRUE, the function will enter debug mode for troubleshooting; defaults to FALSE.}
}
\value{
Numeric, the p-value for the test.
}
\description{
Conducts an exact test for Hardy-Weinberg equilibrium for a given set of genotypic frequencies,
specified by the number of individuals (N), the number of minor alleles (nA), and the observed
number of heterozygotes (nAB), considering a specified inbreeding coefficient (theta).
The function allows for testing whether the observed heterozygosity is significantly less than,
greater than, or different from the expected heterozygosity under Hardy-Weinberg equilibrium.
}
\examples{
N <- 100
K <- 21 # minor allele count
M <- 3 # 3 homozygous carriers
p_value <- hwe_exact_test(N, K, M, theta = 4, alternative = "less")

}
