% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{log_prob_increase}
\alias{log_prob_increase}
\title{Calculate Log Probability Increase for Heterozygotes}
\usage{
log_prob_increase(nAA, nBB, nAB, theta)
}
\arguments{
\item{nAA}{Integer, the number of homozygous minor genotypes.}

\item{nBB}{Integer, the number of homozygous major genotypes.}

\item{nAB}{Integer, the current number of heterozygotes.}

\item{theta}{Numeric, the theta parameter reflecting the relationship between heterozygotes and homozygotes frequencies.}
}
\value{
Numeric, the change in log probability resulting from an increase of two heterozygotes.
}
\description{
Computes the increase in log probability when the number of heterozygotes increases by two,
given the counts of homozygous minor (nAA), homozygous major (nBB) genotypes, current heterozygotes (nAB),
and the theta parameter.
}
