% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_sum_exp.R
\name{log_sum_exp}
\alias{log_sum_exp}
\title{Log-Sum-Exp Trick}
\usage{
log_sum_exp(log_values)
}
\arguments{
\item{log_values}{Numeric vector, log-transformed values to be summed.}
}
\value{
Numeric, the log of the sum of exponentials of the input values.
}
\description{
Performs stable computation of the logarithm of the sum of exponentials of input log values.
This function is particularly useful when dealing with log probabilities to prevent
numerical underflow or overflow. It also handles cases where input values may include `-Inf`.
}
\examples{
log_probs <- c(-Inf, log(0.01), log(0.99))
log_sum <- log_sum_exp(log_probs)
print(log_sum)

}
