% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_sum_probability_of_M_pairs.R
\name{log_sum_probability_of_M_pairs}
\alias{log_sum_probability_of_M_pairs}
\title{Sum Probability of M Pairs}
\usage{
log_sum_probability_of_M_pairs(N, K, M, midp = TRUE)
}
\arguments{
\item{N}{Numeric, the total number of individuals in the population.}

\item{K}{Numeric, the total number of mutated gene copies observed.}

\item{M}{Numeric, the number of samples that have a biallelic variant in the gene.}

\item{midp}{Logical, whether to use the mid-p adjustment in the probability calculation.
The mid-p adjustment is used to correct for discrete probability distributions
in hypothesis testing. Default is TRUE.}
}
\value{
A list containing:
\itemize{
  \item{gt}{Logical, indicating whether the function iterated over greater (TRUE) or
      lesser (FALSE) half of the possible M values.}
  \item{prob}{Numeric vector, the calculated probabilities for each M value.}
  \item{prob_midp}{Numeric vector or NULL, the mid-p adjusted probabilities if `midp` is TRUE,
      otherwise NULL.}
}
}
\description{
This function calculates the sum of probabilities of observing a given number of samples
with a biallelic variant in a gene, assessing deviation from Hardy-Weinberg equilibrium
at the gene level. It is particularly useful for evaluating the genetic structure
and variation within a population.
}
\examples{
# Calculate the sum probability of M pairs for a given N, K, and M, with mid-p adjustment
result <- log_sum_probability_of_M_pairs(N = 100, K = 20, M = 5, midp = TRUE)
print(result)

}
