% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hwe_exact_test.R
\name{hwe_exact_test}
\alias{hwe_exact_test}
\title{Perform an Exact Test under Hardy-Weinberg Equilibrium}
\usage{
hwe_exact_test(N, nA, nAB, theta = 4, alternative = "less", use_mid_p = TRUE)
}
\arguments{
\item{N}{Integer, the total number of diploid individuals in the population.}

\item{nA}{Integer, the total number of copies of the minor allele in the population.}

\item{nAB}{Integer, the observed number of heterozygotes.}

\item{theta}{Numeric, the theta parameter reflecting the inbreeding coefficient; defaults to 4.}

\item{alternative}{Character, specifies the alternative hypothesis and must be one of "less", "greater", or "two.sided".}

\item{use_mid_p}{Logical, indicating whether to use the mid-p correction for the p-value calculation in one-sided tests; defaults to TRUE.}
}
\value{
Numeric, the p-value for the test.
}
\description{
Conducts an exact test for Hardy-Weinberg equilibrium for a given set of genotypic frequencies,
specified by the number of individuals (N), the number of minor alleles (nA), and the observed
number of heterozygotes (nAB), considering a specified inbreeding coefficient (theta).
The function allows for testing whether the observed heterozygosity is significantly less than,
greater than, or different from the expected heterozygosity under Hardy-Weinberg equilibrium.
}
\examples{
N <- 100
nA <- 21
nAB <- 17
p_value <- hwe_exact_test(N, nA, nAB, theta = 4, alternative = "less")

}
